dropped redundant files
